refactor: Remove nested components in components/Login
