+++ src/components/LogIn/IdentityProviderForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./IdentityProviderForm";
