+++ src/components/LogIn/LogIn.test.tsx
@@ -2,9 +2,6 @@ import { screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { vi } from "vitest";
 
-import { thunks as appThunks } from "store/app";
-import { actions as generalActions } from "store/general";
-import * as dashboardStore from "store/store";
 import { configFactory, generalStateFactory } from "testing/factories/general";
 import { rootStateFactory } from "testing/factories/root";
 import { renderComponent } from "testing/utils";
@@ -13,32 +10,8 @@ import LogIn from "./LogIn";
 import { ErrorResponse, Label } from "./types";
 
 describe("LogIn", () => {
-  const consoleError = console.error;
-
-  beforeEach(() => {
-    console.error = vi.fn();
-  });
-
   afterEach(() => {
-    console.error = consoleError;
     vi.restoreAllMocks();
-  });
-
-  it("renders a 'connecting' message while connecting", () => {
-    const state = rootStateFactory.build({
-      general: generalStateFactory.withConfig().build({
-        config: configFactory.build({
-          identityProviderAvailable: true,
-        }),
-      }),
-    });
-    renderComponent(<LogIn>App content</LogIn>, { state });
-    expect(
-      within(screen.getByRole("button")).getByText("Connecting..."),
-    ).toBeInTheDocument();
-    const content = screen.getByText("App content");
-    expect(content).toBeInTheDocument();
-    expect(content).toHaveClass("app-content");
   });
 
   it("does not display the login form if the user is logged in", () => {
@@ -58,38 +31,6 @@ describe("LogIn", () => {
     });
     renderComponent(<LogIn>App content</LogIn>, { state });
     expect(document.querySelector(".login")).not.toBeInTheDocument();
-  });
-
-  it("renders an IdentityProvider login UI if the user is not logged in", () => {
-    const state = rootStateFactory.build({
-      general: generalStateFactory.withConfig().build({
-        visitURLs: ["I am a url"],
-        config: configFactory.build({
-          identityProviderAvailable: true,
-        }),
-      }),
-    });
-    renderComponent(<LogIn>App content</LogIn>, { state });
-    expect(
-      screen.getByRole("link", { name: "Log in to the dashboard" }),
-    ).toBeInTheDocument();
-    const content = screen.getByText("App content");
-    expect(content).toBeInTheDocument();
-    expect(content).toHaveClass("app-content");
-  });
-
-  it("renders a UserPass login UI if the user is not logged in", () => {
-    const state = rootStateFactory.build({
-      general: generalStateFactory.withConfig().build({
-        config: configFactory.build({
-          identityProviderAvailable: false,
-        }),
-      }),
-    });
-    renderComponent(<LogIn>App content</LogIn>, { state });
-    expect(screen.getByRole("button")).toHaveTextContent(
-      "Log in to the dashboard",
-    );
     const content = screen.getByText("App content");
     expect(content).toBeInTheDocument();
     expect(content).toHaveClass("app-content");
@@ -142,50 +83,6 @@ describe("LogIn", () => {
     expect(screen.getByText(Label.INVALID_FIELD)).toBeInTheDocument();
   });
 
-  it("logs in", async () => {
-    // Mock the result of the thunk to be a normal action so that it can be tested
-    // for. This is necessary because we don't have a full store set up which
-    // can dispatch thunks (and we don't need to handle the thunk, just know it
-    // was dispatched).
-    vi.spyOn(appThunks, "connectAndStartPolling").mockImplementation(
-      vi.fn().mockReturnValue({
-        type: "connectAndStartPolling",
-        catch: vi.fn(),
-      }),
-    );
-    const mockUseAppDispatch = vi.fn().mockReturnValue({
-      then: vi.fn().mockReturnValue({ catch: vi.fn() }),
-    });
-    vi.spyOn(dashboardStore, "useAppDispatch").mockReturnValue(
-      mockUseAppDispatch,
-    );
-    const state = rootStateFactory.build({
-      general: generalStateFactory.withConfig().build({
-        config: configFactory.build({
-          identityProviderAvailable: false,
-        }),
-      }),
-    });
-    renderComponent(<LogIn>App content</LogIn>, { state });
-    await userEvent.type(
-      screen.getByRole("textbox", { name: "Username" }),
-      "eggman",
-    );
-    await userEvent.type(screen.getByLabelText("Password"), "verysecure123");
-    await userEvent.click(screen.getByRole("button"));
-    const storeAction = generalActions.storeUserPass({
-      wsControllerURL: "wss://controller.example.com",
-      credential: { user: "eggman", password: "verysecure123" },
-    });
-    expect(mockUseAppDispatch.mock.calls[0][0]).toMatchObject({
-      type: "general/cleanupLoginErrors",
-    });
-    expect(mockUseAppDispatch.mock.calls[1][0]).toMatchObject(storeAction);
-    expect(mockUseAppDispatch.mock.calls[2][0]).toMatchObject({
-      type: "connectAndStartPolling",
-    });
-  });
-
   it("displays authentication request notifications", async () => {
     const state = rootStateFactory.build({
       general: generalStateFactory.withConfig().build({
@@ -223,32 +120,5 @@ describe("LogIn", () => {
       { pointerEventsCheck: 0 },
     );
     expect(screen.queryByTestId("toast-card")).not.toBeInTheDocument();
-  });
-
-  it("should display console error when trying to log in", async () => {
-    vi.spyOn(appThunks, "connectAndStartPolling").mockImplementation(
-      vi.fn().mockReturnValue({ type: "connectAndStartPolling" }),
-    );
-    vi.spyOn(dashboardStore, "useAppDispatch").mockImplementation(
-      vi
-        .fn()
-        .mockReturnValue((action: unknown) =>
-          action instanceof Object &&
-          "type" in action &&
-          action.type === "connectAndStartPolling"
-            ? Promise.reject(
-                new Error("Error while dispatching connectAndStartPolling!"),
-              )
-            : null,
-        ),
-    );
-
-    renderComponent(<LogIn>App content</LogIn>);
-    await userEvent.click(screen.getByRole("button"));
-    expect(appThunks.connectAndStartPolling).toHaveBeenCalledTimes(1);
-    expect(console.error).toHaveBeenCalledWith(
-      Label.POLLING_ERROR,
-      new Error("Error while dispatching connectAndStartPolling!"),
-    );
   });
 });
