+++ src/components/LogIn/UserPassForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UserPassForm";
