+++ src/components/LogIn/index.ts
@@ -1,2 +1,1 @@
 export { default } from "./LogIn";
-export { Label as LogInLabel } from "./types";
